Grading
Excellent work on this and the course overall!

Just a few funky things in your UI, not placing stuff in boxes. Excellent usage of the API's. Would have liked to see you use leafletProxy cause I think you could've handled it but due dates are due dates.

It was a pleasure having you in class this semester, good luck with the rest of your final year of grad school!